# api/openapi.yaml
openapi: 3.0.3
info:
  title: Agora API
  version: 1.0.0
  description: |
    Agora School Management Platform API.
    Base contract aligned with Step 3 (auth, attendance, homework, marks, messaging, notifications).
servers:
  - url: /api/v1
    description: Relative base URL
...
paths:
...
  /files/upload-url:
    post:
      tags: [Files]
      operationId: createSignedUploadUrl
      summary: Generate signed upload URL (and download URL for same object)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FileUploadUrlRequest'
      responses:
        '200':
          description: Signed upload URL generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileUploadUrlResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '422':
          $ref: '#/components/responses/ValidationError'
  /files/download-url:
    post:
      tags: [Files]
      operationId: createSignedDownloadUrl
      summary: Generate signed download URL for existing object key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FileDownloadUrlRequest'
      responses:
        '200':
          description: Signed download URL generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileDownloadUrlResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '422':
          $ref: '#/components/responses/ValidationError'
  /files/local/upload/{token}:
    put:
      tags: [Files]
      operationId: localSignedUpload
      summary: Upload binary payload using local signed token (development)
      security: []
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        '200':
          description: Local upload complete
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileLocalUploadResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '422':
          $ref: '#/components/responses/ValidationError'
  /files/local/download/{token}:
    get:
      tags: [Files]
      operationId: localSignedDownload
      summary: Download binary payload using local signed token (development)
      security: []
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Binary file content
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    DeviceApiKey:
      type: apiKey
      in: header
      name: X-Device-Api-Key
    InternalApiKey:
      type: apiKey
      in: header
      name: X-Internal-Api-Key
...
  responses:
    UnauthorizedError:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ForbiddenError:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFoundError:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    Meta:
      type: object
      properties:
        request_id:
          type: string
          format: uuid
...
    ErrorItem:
      type: object
      properties:
        field:
          type: string
        issue:
          type: string
    ErrorBody:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            $ref: '#/components/schemas/ErrorItem'
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          example: false
        error:
          $ref: '#/components/schemas/ErrorBody'
        meta:
          $ref: '#/components/schemas/Meta'
...
    FileProvider:
      type: string
      enum: [local, s3, gcs]
    SignedFileTransport:
      type: object
      properties:
        provider:
          $ref: '#/components/schemas/FileProvider'
        method:
          type: string
          example: PUT
        url:
          type: string
          format: uri
        headers:
          type: object
          additionalProperties:
            type: string
        expires_in:
          type: integer
          example: 900
    FileUploadUrlRequest:
      type: object
      required: [file_name, content_type]
      properties:
        scope:
          type: string
          enum: [homework, submission, message, profile, general]
        file_name:
          type: string
        content_type:
          type: string
          example: application/pdf
        size_bytes:
          type: integer
          minimum: 1
    FileDownloadUrlRequest:
      type: object
      required: [object_key]
      properties:
        object_key:
          type: string
    FileUploadUrlData:
      type: object
      properties:
        provider:
          $ref: '#/components/schemas/FileProvider'
        bucket:
          type: string
        object_key:
          type: string
        upload:
          $ref: '#/components/schemas/SignedFileTransport'
        download:
          $ref: '#/components/schemas/SignedFileTransport'
    FileDownloadUrlData:
      type: object
      properties:
        provider:
          $ref: '#/components/schemas/FileProvider'
        bucket:
          type: string
        object_key:
          type: string
        download:
          $ref: '#/components/schemas/SignedFileTransport'
    FileUploadUrlResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/FileUploadUrlData'
        meta:
          $ref: '#/components/schemas/Meta'
    FileDownloadUrlResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/FileDownloadUrlData'
        meta:
          $ref: '#/components/schemas/Meta'
    FileLocalUploadResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            uploaded:
              type: boolean
            object_key:
              type: string
            size_bytes:
              type: integer
        meta:
          $ref: '#/components/schemas/Meta'
...
security:
  - BearerAuth: []